apps/api/pyproject.toml
[project]
name = "embedded-lab-api"
version = "0.1.0"
description = "Backend API for Embedded Lab"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0,<1.0.0",
  "uvicorn[standard]>=0.30.0,<1.0.0",
  "pydantic-settings>=2.3.0,<3.0.0",
  "sqlalchemy>=2.0.0,<3.0.0",
  "passlib[bcrypt]>=1.7.4,<2.0.0",
  "bcrypt==4.0.1",
  "python-jose[cryptography]>=3.3.0,<4.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest",
  "httpx"
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
